refactor form and validation
