+++ frontend/src/views/pages/TodosPage/Todos.page.tsx
@@ -1,48 +1,11 @@
 import { AppEnv } from "@/constants/env";
 import { Todo } from "@/modules/todo/type";
-import {
-  Button,
-  FormControl,
-  FormLabel,
-  Input,
-  Text,
-  VStack,
-} from "@chakra-ui/react";
-import React, { useState } from "react";
-
-type FormData = {
-  title: string;
-  content: string;
-};
+import { TodoForm } from "@/views/features/TodoForm";
+import { Button, Text, VStack } from "@chakra-ui/react";
+import { useState } from "react";
 
 export const Todos = () => {
   const [todos, setTodos] = useState<Todo[]>([]);
-  const [formData, setFormData] = useState<FormData>({
-    title: "",
-    content: "",
-  });
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    console.log(JSON.stringify(formData));
-    await fetch(`${AppEnv.apiUrl}/todos`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(formData),
-    });
-    setFormData({ title: "", content: "" });
-  };
-
-  const handleInputChange = (
-    e: React.ChangeEvent<
-      HTMLInputElement | HTMLTextAreaElement | HTMLSelectElement
-    >
-  ) => {
-    const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value });
-  };
 
   return (
     <>
@@ -77,31 +40,7 @@ export const Todos = () => {
             </div>
           );
         })}
-        <form onSubmit={handleSubmit}>
-          <FormControl mt={3}>
-            <FormLabel>Title</FormLabel>
-            <Input
-              type="text"
-              id="title"
-              name="title"
-              value={formData.title}
-              onChange={handleInputChange}
-            />
-          </FormControl>
-          <FormControl mt={3}>
-            <FormLabel>Content</FormLabel>
-            <Input
-              type="text"
-              id="content"
-              name="content"
-              value={formData.content}
-              onChange={handleInputChange}
-            />
-          </FormControl>
-          <Button type="submit" mt={3} justifyContent={"center"}>
-            Create Todo
-          </Button>
-        </form>
+        <TodoForm />
       </VStack>
     </>
   );
